==> code/00_prep.Rmd <==
---
title: "Prepare BUNMD data"
---

Summary: This script cleans the BUNMD data and creates a unique key for households (by concatenating parents' names). This does nicknames and bni and outputs data set for making figures and modeling subsetting: only males, only blacks and whites note. No restrictions on byear. 

```{r}
library(data.table)
library(tidyverse)
library(RecordLinkage)
```

## 1 Read in BUNMD 

```{r}
dt <- fread("/censoc/data/censoc_v2/bunmd_v2.csv")

dt <- dt %>% 
  filter(!is.na(sex)) %>% 
  filter(!is.na(race_first))

## rename race 
dt[, race := race_first]

## drop everyone whose not black or white
dt = dt[race %in% 1:2]

## 1c. keep only males
dt = dt[sex == 1]

initial_size = nrow(dt)
print(initial_size)
```

## 2 Clean names

In this section, we clean names. 

```{r}
## First names cleaning
## various NA strings
na.strings = c("^$", # empty string
               "^NA$",
               ## varieties of "UNKNOWN"
               "^UKN", # start with UKN ...
               "^UNK", # start with UNK ...
               "^UN$", "^U$", "^NS$",
               "^UK$",
               ## varieties of NOT and NONE
               "NONENAMED",
               "^NONE$",
               "^NOT",
               "NONAME",
               ## variaeties of X, XX, XXXXXX
               "^X*$",
               ## other
               "UNSTATED",
               "UNLISTED",
               "^UNAV", ## unavailable and misspellings thereof
               "MISSING",
               "^UNDISCLO",
               "^UNSPEC",
               "WITHHELD",
               "DECEASED",
               "^NAME$")
na.regx = paste(na.strings, collapse = "|") ## regexp for 'or'
clean_first_names <- function(fname)
{
  ## Missing --> NA
  fname[grepl(pattern = na.regx, fname)] <- NA
  ## first word only
  fname <- gsub(pattern = " .*$", replacement = "", x = fname)
  ## NA if 0 or 1 characters
  fname[nchar(fname) < 2] <- NA
  fname
}

## carry out
dt[, fname_clean := clean_first_names(fname)]
dt[, mother_fname_clean := clean_first_names(mother_fname)]
dt[, father_fname_clean := clean_first_names(father_fname)]
dt[, father_lname_clean := clean_first_names(father_lname)]
dt[, mother_lname_clean := clean_first_names(mother_lname)]
```

## 3 Get family (keys)

```{r}
dt[!is.na(father_lname) & father_lname != "" &
     !is.na(father_fname_clean) &
     !is.na(mother_lname) & mother_lname != "" &
     !is.na(mother_fname_clean) ,
   family := paste(father_lname_clean,
                   father_fname_clean,
                   mother_lname_clean,
                   mother_fname_clean, sep = "_")]
## if family == "" --> NA
dt[family == "", family := NA]
```

## 4 Standardize names using IPUMS dictionary

```{r}
## read in mpc + abe nickname file (cleaned)
nickname_file <- fread("../data/nickname_crosswalk_master.csv")

## josh's named vector trick
## we have a vector with observed names as labels and standardized names as values
male_nickname.vec = nickname_file[sex == 1]$fname_std
names(male_nickname.vec) = nickname_file[sex == 1]$fname
## tail(male_nickname.vec)
##    ROYCE   SUMMERS   WALDRON        ED     EDDIE      MOSE
##  "ROYCE" "SUMMERS" "WALDRON"  "EDWARD"  "EDWARD"   "MOSES"
male_nickname.vec[male_nickname.vec == "WILLIAM"]
##   WILLIAM        WM    WILLIE      WILL      BILL     WILLY   WILLIAN     BILLY
## "WILLIAM" "WILLIAM" "WILLIAM" "WILLIAM" "WILLIAM" "WILLIAM" "WILLIAM" "WILLIAM"
##    WILIAM    BILLIE      WILY    WILLAM     WILLI    WILLIA     WILLE     WILLM

dt[, fname_std := fname_clean] ## we default to original clean name
dt[fname_clean %in% names(male_nickname.vec), ## if appears in nickname file, standardize
   fname_std := male_nickname.vec[fname_clean] ]
## dt[, fname_std := male_nickname.vec[fname_clean] ]
```

## 5 Generate BNI estimates (standardized names)

```{r}
## no restriction on byear
tt = dt[race %in% 1:2,
        table(fname_std, race)]
ptt = prop.table(tt, 2)
p_black = ptt[,2]
p_white = ptt[,1]
bni.dt = data.table(fname_std = rownames(tt),
                    bni = p_black/ (p_black + p_white),
                    N = rowSums(tt))
## use same named-vector trick for assigning bni to dt
bni.vec = bni.dt$bni
names(bni.vec) = bni.dt$fname_std
dt[, bni_std := bni.vec[fname_std]]

## use same named-vector trick for assigning proportion black to dt
prop.dt = data.table(fname_std = rownames(tt),
                    prop_black = tt[,2]/(tt[,1] + tt[,2]),
                    N = rowSums(tt))

prop.vec = prop.dt$prop_black
names(prop.vec) = prop.dt$fname_std
dt[, prop_black := prop.vec[fname_std]]


## write out bni crosswalk file 
fwrite(x = bni.dt %>% 
  filter(N > 10), file = "../data/bni_crosswalk.csv")


```

## 5.1 Generate BNI estimates (unstandardized names)

```{r}
## no restriction on byear
tt = dt[race %in% 1:2,
        table(fname_clean, race)]
ptt = prop.table(tt, 2)
p_black = ptt[,2]
p_white = ptt[,1]
bni.dt = data.table(fname_clean = rownames(tt),
                    bni = p_black/ (p_black + p_white),
                    N = rowSums(tt))
## use same named-vector trick for assigning bni to dt
bni.vec = bni.dt$bni
names(bni.vec) = bni.dt$fname_clean
dt[, bni_unstandardized := bni.vec[fname_clean]]

## use same named-vector trick for assigning proportion black to dt
prop.dt = data.table(fname_clean = rownames(tt),
                    prop_black = tt[,2]/(tt[,1] + tt[,2]),
                    N = rowSums(tt))

prop.vec = prop.dt$prop_black
names(prop.vec) = prop.dt$fname_clean
dt[, prop_black_unstandardized := prop.vec[fname_clean]]
```


## 6 Add Social Security Benefits 

```{r}
## read in social security benefit level
oasdi <- fread("../data/archive/oasdi_zip05_combined_new.csv")

oasdi_small <- oasdi[, .(zip, zip_ben)]
oasdi_small[, .N, by = zip]

## merge data
dt[, zip5 := as.numeric(substr(zip_residence, 1, 5))]

dt <- merge(dt, oasdi_small[!is.na(zip)], by.x = "zip5", by.y = "zip", all.x = T)
```

## 7 Export data set

```{r}
## drop a few columns that we really don't need
columns_to_drop = c("race_first_cyear",
                    "race_last_cmonth",
                    "race_first_cmonth",
                    "number_claims") 
## https://stackoverflow.com/questions/9202413/how-do-you-delete-a-column-by-name-in-data-table
dt[, (columns_to_drop) := NULL]

## add a column with fname_std_freq
dt[, fname_std_freq := .N, by = fname_std]


fwrite(x = dt, file = "/censoc/data/working_files/00_cleaned_sibs.csv")
```

## 8 Identify additional siblings  

```{r}
## Filter to Black Males with non-missing parents names
potential_sibs <- dt %>%
  filter(sex == 1 & race == 2) %>% 
  filter(!(is.na(father_fname_clean) | father_fname_clean == "")) %>% 
  filter(!(is.na(father_lname_clean) | father_lname_clean == "")) %>% 
  filter(!(is.na(mother_fname_clean) | mother_fname_clean == "")) %>% 
  filter(!(is.na(mother_lname_clean) | mother_lname_clean == "")) %>% 
  filter(!(is.na(fname_std) | fname_std == "")) 

## Restrict to men without sibling already identified 
potential_sibs <- potential_sibs %>% 
  group_by(family) %>% 
  filter(n() == 1)

# Establish siblingship exact match + similiar father's first name  ----------------

## Create keys for sibs who match on bpl, father lname, mother lname, father fname, but differ on mother fname
additional_sibs_mother_mismatch <- potential_sibs %>% 
  mutate(family = paste(father_lname_clean, mother_lname_clean, father_fname_clean, "MOTHER-FNAME-STD", bpl, sep = "_")) 

## Compare mother fname for potential sibs; establish as sibs if lev_distance > 0.7
new_sibs_mother <- additional_sibs_mother_mismatch %>%
  group_by(family) %>% 
  filter(n() == 2) %>% 
  mutate(lev_distance = levenshteinSim(mother_fname_clean[1], mother_fname_clean[2]))  %>% 
  filter(lev_distance > .7) %>% 
  arrange(family)

# Establish siblingship exact match + similiar mother's first name  ----------------

## find potential sibs who match on bpl, father lname, mother lname, father fname, but differ on father fname
additional_sibs_father_mismatch <- potential_sibs %>% 
  mutate(family = paste(father_lname_clean, mother_lname_clean, mother_fname_clean, "FATHER-FNAME-STD", bpl, sep = "_")) 

## Compare father fname for potential sibs; establish as sibs if lev_distance > 0.7
new_sibs_father <- additional_sibs_father_mismatch %>%
  group_by(family) %>% 
  filter(n() == 2) %>% 
  mutate(lev_distance_father = levenshteinSim(father_fname_clean[1], father_fname_clean[2])) %>% 
  filter(lev_distance_father > .7) %>% 
  arrange(family)

# establish siblingship for pairs with similar (but not identical) mother *and* father first name --------

## find potential sibs who match on bpl, father lname, mother lname, but differ on mother fname and father fname 
additional_sibs_fnames_mismatch <- potential_sibs %>% 
  mutate(family = paste(father_lname_clean, mother_lname_clean, "FNAMES-STD", bpl, sep = "_")) 

## Compare father fname for potential sibs; establish as sibs if lev_distance > 0.8
new_sibs_mother_father_mismatch <- additional_sibs_fnames_mismatch %>%
  group_by(family) %>% 
  filter(n() == 2) %>% 
  mutate(lev_distance_mother = levenshteinSim(mother_fname_clean[1], mother_fname_clean[2]),
         lev_distance_father = levenshteinSim(father_fname_clean[1], father_fname_clean[2]))  %>% 
  filter(lev_distance_mother > .8 & lev_distance_father > .8) 

## select sibs not already matched 
new_sibs_mother_father <- new_sibs_mother_father_mismatch %>% 
  filter(!(lev_distance_mother == 1 | lev_distance_father == 1)) 

## combine all sibs
new_sibs <- bind_rows(new_sibs_mother, new_sibs_father, new_sibs_mother_father) %>% 
  arrange(family) %>% 
  mutate(exact_match = 0)
```

## 9 Create Analysis File

```{r}
## create analysis file 
all_black_sibs <- dt %>% 
  filter(race == 2) %>% 
  group_by(family) %>% 
  filter(n() %in% 2:5) %>% 
  mutate(exact_match = 1)

analysis_file <- bind_rows(all_black_sibs, new_sibs)

## calculate (relative) birth order 
analysis_file <- analysis_file %>% 
  group_by(family) %>%
  arrange(byear, bmonth, bday, .by_group = TRUE) %>%
  mutate(birth_order = row_number())

## write out analysis file 
fwrite(x = analysis_file, file = "../data/analysis_file.csv")
```
