==> code/01_tables.Rmd <==
---
title: "Analysis"
---

Summary: Analysis for black lives longevity project. 

## Setup 

Read in package 
```{r setup}
## library packages
library(data.table)
library(tidyverse)
library(stargazer)
library(broom) ## puts model output into "tidy" data.frame
library(cowplot) ## ggplot themes 
library(lfe) ## deprecated on cran
library(huxtable) ## output regression tables 
library(fixest) ## fast fixed-effect linear models
library(here) ## relative paths
library(ggpubr)

## set threads for computing 
setFixest_nthreads(1)

siblings <- fread(here("data/analysis_file.csv")) %>%  filter(bpl < 15000)

siblings_restricted <- fread(here("data/analysis_file_restricted.csv")) %>%  filter(bpl < 15000)
```


## Main Models (for Black)

First, we run an analysis with fixed effects for birth cohort, family, and birth order. 

```{r}
## siblings with no limits on death year imposed 
all_sibs_nolim <- siblings %>% 
  filter(byear %in% 1915:1925)  %>% 
  filter(!is.na(bni_std)) %>% 
  filter(fname_std_freq > 500) %>% 
  group_by(family) %>% 
  filter(n() %in% 2:5) 

model = feols(death_age ~ bni_std | byear, data = siblings_restricted) 
model_family_byear = feols(death_age ~ bni_std | byear + family ,  data = siblings_restricted)
model_family_byear_birth_order = feols(death_age ~ bni_std | byear + family + birth_order ,  data = siblings_restricted)
model_family_byear_birth_order_nolim = feols(death_age ~ bni_std | byear + family + birth_order,  data = all_sibs_nolim)

etable(model, model_family_byear, model_family_byear_birth_order, model_family_byear_birth_order_nolim, tex = T) # option to save: , file = here("tables/table1_main_models.tex")
```

```{r}
model_family_byear = feols(death_age ~ bni_std | byear + family ,  data = siblings_restricted)

model_family_byear_family = feols(death_age ~ bni_std |  byear + family  ,  data = siblings_restricted)

summary(lfe::felm(death_age ~ bni_std | byear + family ,  data = siblings_restricted))

```


## Black vs. White Sibs

```{r, eval = F}
## all sibs 
bunmd_sibs <- fread("/censoc/data/working_files/00_cleaned_sibs.csv")

white_sibs <- bunmd_sibs %>% 
  filter(bpl < 15000) %>% 
  filter(byear %in% 1915:1925)  %>% 
  filter(dyear %in% 1988:2005) %>% 
  filter(!is.na(bni_std)) %>% 
  filter(fname_std_freq > 500) %>% 
  filter(race == 1) %>% 
  group_by(family) %>% 
  filter(n() %in% 2:5)

## birth order 
white_sibs <- white_sibs %>% 
  group_by(family) %>%
  arrange(byear, bmonth, bday, .by_group = TRUE) %>%
  mutate(birth_order = row_number())

model.whites = feols(death_age ~ bni_std | byear + family + birth_order,  data = white_sibs, nthreads = 1)

model.blacks = feols(death_age ~ bni_std | byear + family + birth_order, data = all_sibs, nthreads = 1)

etable(model.whites, model.blacks, tex = T)
```

## Zip Benefits (Outcome)

Rerun model with Zip an outcome variable. Do people with blacker names end up living in richer zip codes?

```{r Zip Benefits Analysis}
model_byear = feols(zip_ben ~ bni_std | byear,  data = siblings_restricted)
model_family_byear = feols(zip_ben ~ bni_std | byear + family,  data = siblings_restricted)
model_family_byear_birth_order = feols(zip_ben ~ bni_std | birth_order + family + byear,  data = siblings_restricted)

etable(model, model_family_byear, model_family_byear_birth_order) # tex = T, file = here("tables/table3_main_models.tex")
```


# Appendix 

## exact match siblings 
```{r}
## define exact sibs dataset 
additional_sibs <- siblings_restricted %>% 
  filter(exact_match == 0) 
 
## define additional sibs 
additional_sibs <- siblings_restricted %>% 
  filter(exact_match == 0) 

## define all sibs 
all_sibs <- siblings_restricted

bni.add.sibs = feols(death_age ~ bni_std | byear + birth_order + family ,  data = additional_sibs)
bni.exact.sibs = feols(death_age ~ bni_std | byear + birth_order + family ,  data = exact_sibs)
bni.all.sibs = feols(death_age ~ bni_std | byear + birth_order + family ,  data = all_sibs)

etable(bni.exact.sibs, bni.add.sibs, bni.all.sibs, tex = T)
```


## Table: Establishing siblingship  

The code below outputs two tables providing concrete examples of our procedure for identifying siblings in this analysis. 

```{r create tables for siblingship}
## create table of matches 
knitr::kable(siblings %>% 
  filter(family %in% sample(unique(family), 3)) %>% 
  select(fname_std, lname, father_fname_clean, father_lname_clean, mother_fname_clean, mother_lname_clean, bpl_string) %>%
  mutate(across(where(is.character), tolower)) %>% 
  mutate(across(where(is.character), tools::toTitleCase)),
  format = "latex", booktabs = T)  

## create table of matches 
knitr::kable(siblings %>% 
  filter(lev_distance_father > 0) %>% 
  filter(family %in% sample(unique(family), 3)) %>% 
  select(fname_std, lname, bpl_string, father_fname_clean, father_lname_clean, mother_fname_clean, mother_lname_clean, lev_distance_father) %>% 
  mutate(across(where(is.character), tolower)) %>% 
  mutate(across(where(is.character), tools::toTitleCase)),
  format = "latex", booktabs = T)  
```

## Miscallaneous Analyses   


```{r}
model_family = feols(death_age ~ prop_black | family,  data = all_sibs)
model_family_byear = feols(death_age ~ prop_black | family  + byear,  data = all_sibs)
model_family_byear_birth_order = feols(death_age ~ prop_black | family + birth_order + byear,  data = all_sibs)

etable(model_family, model_family_byear, model_family_byear_birth_order, tex = T)
```


## Warren Analysis

```{r Warren Analysis}
## same birth cohorts (1910-1920) 
## unstandardized names, predictor is percentage black  
## 

# warren_replication <- siblings %>% 
#   filter(byear %in% 1910:1920) %>% 
#   filter(!is.na(prop_black_unstandardized)) %>% 
#   filter(fname_unstandardized_freq > 100) %>% 
#   group_by(family) %>% 
#   filter(n() %in% 2:5) 
# 
# 
# warren_model = feols(death_age ~ prop_black_unstandardized | family + birth_order,  data = warren_replication)

```



```{r}
birth_order_plot <- tidy(feols(bni_std ~ as.factor(birth_order) | byear, data = siblings_restricted)) %>% 
  mutate(term = str_sub(term, 23, 30)) %>% 
  ggplot(aes(x = term, y = estimate, ymin = estimate - 1.96*std.error, ymax = 1.96*std.error)) + 
  geom_pointrange() +
  theme_cowplot() + 
  geom_hline(yintercept = 0, linetype = "dashed") + 
  labs(x = "Birth Order (Ref = 1)",
       title = "Birth Order")

ggsave(birth_order_plot, filename = here("figures/birth_order.pdf"), height = 4, width = 6)
```

```{r}
siblings_restricted %>% 
  group_by(family) %>% 
  summarize(death_age = mean(death_age), bni = mean(bni_std)) %>% 
  mutate(bni_decile = ntile(bni, 10)) %>% 
  group_by(bni_decile) %>% 
  summarize(death_age_avg = mean(death_age),
            se = sd(death_age) / sqrt(n()))  %>% 
  ggplot(aes(x = bni_decile,
             y = death_age_avg,
             ymin = death_age_avg - 1.96*se,
             ymax = death_age_avg + 1.96*se)) + 
  geom_pointrange() + 
  theme_cowplot() + 
  ylim(74, 77)


family_avg = siblings_restricted %>% 
  group_by(family) %>% 
  summarize(death_age = mean(death_age), bni = mean(bni_std), byear = mean(byear))


feols(death_age ~ bni | byear, data = family_avg)


sib_correlation <- siblings_restricted %>% 
  group_by(family) %>% 
  summarize(death_age = mean(death_age), bni = mean(bni_std)) %>% 
  ggplot(aes(x = bni, y = death_age)) + 
  geom_point() + 
  geom_smooth(method = "lm", se = F) + 
  stat_cor(aes(label = paste(..r.label.., ..p.label.., sep = "~`,`~"))) + 
  theme_cowplot() + 
  labs(x = "Household BNI (All Brothers",
       y = "Death Age",
       title = "Household-level Correlation: BNI vs. Death Age")

ggsave(sib_correlation, filename = here("figures/household_correlation.png"), height = 5, width = 7)

```
